# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Testhub API
  version: 0.6.0
  description: |
    Core API for Testhub (runs, results ingestion).

    Authentication:
    - Protected endpoints require an API key via the `x-api-key` header.
    - Authorization is organization-scoped; non-owned resources return 404.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
...
paths:
...
  /projects/{projectId}/tests:
    get:
      tags: [Tests]
      operationId: listTests
      summary: List test cases for a project
      description: Returns test cases ordered by last-seen result desc.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/TestNameQuery'
        - $ref: '#/components/parameters/TestSuiteFilter'
        - $ref: '#/components/parameters/TestStatusFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/tests/{testCaseId}/history:
    get:
      tags: [Tests]
      operationId: getTestHistory
      summary: Get execution history for a test case
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TestCaseId'
        - $ref: '#/components/parameters/HistoryLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Analytics ----------
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authenticating requests.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Project slug or database id (implementation accepts both).
      schema:
        type: string
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    TestCaseId:
      name: testCaseId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    TestNameQuery:
      name: q
      in: query
      required: false
      description: Filter test cases by name or externalId (substring match).
      schema:
        type: string
    TestSuiteFilter:
      name: suite
      in: query
      required: false
      description: Filter test cases by suiteName (substring match).
      schema:
        type: string
    TestStatusFilter:
      name: status
      in: query
      required: false
      description: Filter test cases by last-seen status.
      schema:
        $ref: '#/components/schemas/TestStatus'
    HistoryLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    Unauthorized:
      description: Unauthorized (missing or invalid API key)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                statusCode: 401
                error: Unauthorized
                message: Authentication required
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                statusCode: 404
                error: Not Found
                message: Not found
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation error

    # ---------- Projects ----------
...
    RunStatus:
      type: string
      enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELED]
    TestStatus:
      type: string
      enum: [PASSED, FAILED, SKIPPED, ERROR]
    NullableTestStatus:
      type: string
      enum: [PASSED, FAILED, SKIPPED, ERROR]
      nullable: true
...
    TestCaseListItem:
      type: object
      required: [id, externalId, name, tags, createdAt, lastStatus, lastSeenAt]
      properties:
        id:
          type: string
        externalId:
          type: string
        name:
          type: string
        suiteName:
          type: string
          nullable: true
        filePath:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        lastStatus:
          $ref: '#/components/schemas/NullableTestStatus'
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    TestCaseListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseListItem'
      additionalProperties: false
    TestCaseHistoryItem:
      type: object
      required: [id, status, durationMs, createdAt, run]
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TestStatus'
        durationMs:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        run:
          type: object
          required: [id, createdAt, status, branch, commitSha]
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            status:
              $ref: '#/components/schemas/RunStatus'
            branch:
              type: string
              nullable: true
            commitSha:
              type: string
              nullable: true
          additionalProperties: false
      additionalProperties: false
    TestCaseHistoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseHistoryItem'
      additionalProperties: false
...